ChecksumString type and usage for checksum
In the current metadata it's valid to submit xml files with a checksum attribute that has an empty string as a value. This is not the intended purpose of the attribute. I suggest solving this by introducing a custom type that enforces the value to look like a checksum.

Here the value is required to be 32 or 64 hex characters. 32 for md5 and 64 for sha256. In this proposal capital letters are not allowed.

I'm not sure if there's some special place for custom types or if it's ok to leave them here.